FLoat

+++ cudaApplication/src/cuda_rt.cuh
@@ -3,6 +3,6 @@
 
 #include "utils/cuda_rt_util.cuh"
 
-double* h_raytrace(std::vector<Eigen::Vector3d> ray_origins, std::vector<Eigen::Vector3d> ray_directions, std::vector<Mesh> meshes, int width, int height, std::vector<Eigen::Vector3d> light_positions, std::vector<Eigen::Vector4d> light_colors);
+float* h_raytrace(std::vector<Eigen::Vector3f> ray_origins, std::vector<Eigen::Vector3f> ray_directions, std::vector<Mesh> meshes, int width, int height, std::vector<Eigen::Vector3f> light_positions, std::vector<Eigen::Vector4f> light_colors);
 
 #endif
